requirements.txt
torch
transformers
diffusers
Pillow
tqdm
tkinter